Extract upsell-related logic into a separate logic to avoid loading data unless needed
