+++ frontend/src/models/groupsModel.ts
@@ -1,35 +1,15 @@
 import { kea } from 'kea'
 import api from 'lib/api'
-import { AvailableFeature, GroupType } from '~/types'
+import { GroupType } from '~/types'
 import { teamLogic } from 'scenes/teamLogic'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { FEATURE_FLAGS } from 'lib/constants'
 import { groupsModelType } from './groupsModelType'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
-import { preflightLogic } from 'scenes/PreflightCheck/logic'
-import { userLogic } from 'scenes/userLogic'
+import { groupsAccessLogic } from 'lib/introductions/groupsAccessLogic'
 
-export enum GroupsAccessStatus {
-    AlreadyUsing,
-    HasAccess,
-    HasGroupTypes,
-    NoAccess,
-    Hidden,
-}
-
-export const groupsModel = kea<groupsModelType<GroupsAccessStatus>>({
+export const groupsModel = kea<groupsModelType>({
     path: ['models', 'groupsModel'],
     connect: {
-        values: [
-            teamLogic,
-            ['currentTeam', 'currentTeamId'],
-            featureFlagLogic,
-            ['featureFlags'],
-            preflightLogic,
-            ['clickhouseEnabled', 'preflight'],
-            userLogic,
-            ['hasAvailableFeature'],
-        ],
+        values: [teamLogic, ['currentTeamId'], groupsAccessLogic, ['groupsEnabled']],
     },
     loaders: ({ values }) => ({
         groupTypes: [
@@ -45,36 +25,6 @@ export const groupsModel = kea<groupsModelType<GroupsAccessStatus>>({
         ],
     }),
     selectors: {
-        groupsCanBeEnabled: [
-            (s) => [s.featureFlags, s.clickhouseEnabled],
-            (featureFlags, clickhouseEnabled) => featureFlags[FEATURE_FLAGS.GROUP_ANALYTICS] && clickhouseEnabled,
-        ],
-        groupsEnabled: [
-            (s) => [s.groupsCanBeEnabled, s.hasAvailableFeature],
-            (groupsCanBeEnabled, hasAvailableFeature) =>
-                groupsCanBeEnabled && hasAvailableFeature(AvailableFeature.CORRELATION_ANALYSIS),
-        ],
-        // Used to toggle various upsell mechanisms for groups
-        groupsAccessStatus: [
-            (s) => [s.groupsCanBeEnabled, s.groupsEnabled, s.currentTeam, s.preflight],
-            (canBeEnabled, isEnabled, currentTeam, preflight): GroupsAccessStatus => {
-                const hasGroups = currentTeam?.has_group_types
-                const hideUpsell = preflight?.instance_preferences?.disable_paid_fs
-                if (!canBeEnabled) {
-                    return GroupsAccessStatus.Hidden
-                } else if (isEnabled && hasGroups) {
-                    return GroupsAccessStatus.AlreadyUsing
-                } else if (hideUpsell) {
-                    return GroupsAccessStatus.Hidden
-                } else if (isEnabled) {
-                    return GroupsAccessStatus.HasAccess
-                } else if (hasGroups) {
-                    return GroupsAccessStatus.HasGroupTypes
-                } else {
-                    return GroupsAccessStatus.NoAccess
-                }
-            },
-        ],
         showGroupsOptions: [
             (s) => [s.groupsEnabled, s.groupTypes],
             (enabled, groupTypes) => enabled && groupTypes.length > 0,
