+++ frontend/src/scenes/groups/Groups.tsx
@@ -10,8 +10,8 @@ import { LemonTable } from 'lib/components/LemonTable'
 import { Link } from 'lib/components/Link'
 import { urls } from 'scenes/urls'
 import { SceneExport } from 'scenes/sceneTypes'
-import { GroupsAccessStatus, groupsModel } from '~/models/groupsModel'
 import { GroupsIntroduction } from 'scenes/groups/GroupsIntroduction'
+import { groupsAccessLogic, GroupsAccessStatus } from 'lib/introductions/groupsAccessLogic'
 
 export const scene: SceneExport = {
     component: Groups,
@@ -21,7 +21,7 @@ export const scene: SceneExport = {
 export function Groups(): JSX.Element {
     const { groups, groupsLoading } = useValues(groupsListLogic)
     const { loadGroups } = useActions(groupsListLogic)
-    const { groupsAccessStatus } = useValues(groupsModel)
+    const { groupsAccessStatus } = useValues(groupsAccessLogic)
 
     if (
         groupsAccessStatus == GroupsAccessStatus.HasAccess ||
